agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14.2"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "logfire>=4.16.0",
    "pydantic>=2.12.5",
    "pydantic-ai>=1.39.0",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.17",
    "uvicorn>=0.37.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.10",
    "taskipy>=1.14.1",
    "ty>=0.0.8",
]

[tool.taskipy.tasks]
dev = "ENV_FILE=.env uv run uvicorn app.main:app --host 0.0.0.0 --port 8000 --reload"
lint = "uv run ruff check"
start = "uv run uvicorn app.main:app --host 0.0.0.0 --port 8000"
test = "uv run pytest"
